+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current (in progress)
 
 - Fix: return the correct error when no `Content-Type` is sent instead of 500 [#2967](https://github.com/opendatateam/udata/pull/2967)
+- Improve documentation for API errors [#2952](https://github.com/opendatateam/udata/pull/2965)
 
 ## 7.0.4 (2024-02-27)